+++ app/src/main/java/knaufdan/android/simpletimerapp/databinding/ProgressBarBindingAdapters.kt
@@ -1,6 +1,5 @@
 package knaufdan.android.simpletimerapp.databinding
 
-import android.widget.NumberPicker
 import android.widget.TextView
 import androidx.databinding.BindingAdapter
 import knaufdan.android.simpletimerapp.util.determineClockSections
@@ -15,16 +14,3 @@ fun TextView.setProgressText(progress: Int?) {
 
 private fun Int.addZero() = if (this < 10) "0$this" else this.toString()
 
-@BindingAdapter(value = ["min", "max", "isWrapped"])
-fun NumberPicker.setup(
-    min: Int,
-    max: Int,
-    isWrapped: Boolean
-) {
-    minValue = min
-    maxValue = if (max < min) min else max
-    wrapSelectorWheel = isWrapped
-}
-
-@BindingAdapter(value = ["suffix"])
-fun NumberPicker.configureFormatter(suffix: String) = setFormatter { value -> "$value $suffix" }
