+++ app/src/main/java/knaufdan/android/simpletimerapp/ui/fragments/InputFragmentViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import knaufdan.android.core.SharedPrefService
 import javax.inject.Inject
 import knaufdan.android.simpletimerapp.arch.BaseViewModel
-import knaufdan.android.simpletimerapp.databinding.bindTo
+import knaufdan.android.core.databinding.bindTo
 import knaufdan.android.simpletimerapp.ui.data.TimerConfiguration
 import knaufdan.android.simpletimerapp.ui.navigation.Navigator
 import knaufdan.android.simpletimerapp.util.Constants.HOUR_IN_MILLIS
